[train] mmd

+++ TrainAndTest/common_func.py
@@ -595,7 +595,7 @@ def args_parse(description: str = 'PyTorch Speaker Recognition: Classification')
     # args for a-softmax
     parser.add_argument('--all-iteraion', type=int, default=0, metavar='M',
                         help='the margin value for the angualr softmax loss function (default: 3.0')
-    parser.add_argument('--m', type=int, default=3, metavar='M',
+    parser.add_argument('--m', type=float, default=3, metavar='M',
                         help='the margin value for the angualr softmax loss function (default: 3.0')
     parser.add_argument('--lambda-min', type=int, default=5, metavar='S',
                         help='random seed (default: 0)')
